Code comments to make understanding the big picture easier
I was wondering how we could enhance gocat to support more deployment tools
like helm w/ or w/o gitops.

To figure out how to do this, I wanted to understand the big picture of how
gocat works.  I started by reading the code, but that was a bit overwhelming.
So I decided to add comments to the code to help me understand it better.

Hope this helps you and others too!