Rename DeployProject FilterRegexp -> ImageTagRegexp
so that it is clearer that it is used to filter image tags only. Peviously it was unclear about what are filtered by the regexp. It is image tags that are filtered, so make sure the name clearly represents that for readability.

Additionally, this makes the variable naming consistent between ImageTagRegexp and the existing ImageTagVars. ImageTagVars are used as the template parameters when parsing ImageTagRegexp as a Go template. A consistent naming makes correlation between the two more apparent, which will also help readability.