GitOps plugin system and a "kanvas" plugin placeholder
This is the first step to add the support for kanvas as our alternative deployment tool usable by gocat.

We implement the interactor and the model for kanvas almost reusing our existing InteractorKustomize and ModelKustomize.

The differences between the kustomize and the kanvas supports are the "Prepare" function, which is used to edit the gitops config file in a Git repository and submit a pull request.

The kustomize support does so by using gocat-builtin kustomize support. The new kanvas support does so by calling kanvas.

In this change, we focus on refactoring and adding the placeholder for the kanvas support.

Once this is merged, we will keep adding the actual implementation, gradually replacing the placeholder to a full-fledged feature.